requirements.txt
flask
gunicorn
numpy
opencv-python
tensorflow-cpu
